requirements.txt
albumentations
pytorch
pyyaml
numpy
tifffile
skimage
sklearn
tqdm
einops